Move directory cleanup to cleanup module
